ajout des capteurs ToF dans le code bas niveau

+++ programmation/sumobot/Path.h
@@ -52,8 +52,8 @@
 */
 struct UnitMove
 {
-	int16_t length; // Longueur du mouvement, en ticks pour MotionControlSystem et en mm pour le haut niveau
-	int16_t bendRadius; // Rayon de courbure, en ticks pour MotionControlSystem et en mm pour le haut niveau
+	int32_t length; // Longueur du mouvement, en ticks pour MotionControlSystem et en mm pour le haut niveau
+	int32_t bendRadius; // Rayon de courbure, en ticks pour MotionControlSystem et en mm pour le haut niveau
 	int32_t speed;	// Vitesse de déplacement (POSITIVE), en ticks/s pour MotionControlSystem et en mm/s pour le haut niveau
 };
